refactor: moving signup form
